Move errors from WebSocket upgrade
